requirements.txt
streamlit
ultralytics
opencv-python
pandas
numpy
pillow
PyYAML
pyodbc
